DuoFinder-backend/requirements.txt
fastapi==0.110.2
uvicorn==0.30.0
sqlalchemy==2.0.30
pyodbc==5.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
email-validator==2.4.0
python-multipart==0.0.6
pydantic[email]==2.6.3
